api/requirements.txt
flask
Flask-Cors
gunicorn
pymongo
pydantic
pillow
numpy<2
tflite-runtime==2.14.0
